# DRI_CA/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DRI & CA — Disaster Resilience Intelligence API
  description: |
    Backend API for the Disaster Resilience Intelligence & Community Awareness module.
    Provides spatial feasibility checks (PostGIS), explainable AI remediation recommendations,
    semantic jargon simplification, Bhashini NMT/TTS translation, community alerts, and
    seasonal disaster preparedness tips.

    **Key Features:**
    - PostGIS spatial intersection queries against KSDMA flood and NCESS landslide hazard zones
    - Explainable AI (XAI) engine with transparent reasoning chains and source citations
    - Deterministic jargon simplification (no LLM dependency)
    - Bhashini Neural Machine Translation (English ↔ Malayalam) with TTS
    - Community-sourced disaster alerts with verification workflow
    - Seasonal awareness tips from KSDMA guidelines

    **Tech Stack:** Node.js (ESM) · Express · PostgreSQL + PostGIS
  version: 1.0.0
  contact:
    name: DRI & CA Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/v1/feasibility:
    post:
      tags: [Feasibility]
      summary: Run site feasibility check
      description: |
        Accepts GPS coordinates and building type. Runs PostGIS ST_Intersects queries
        against flood and landslide hazard zones. Returns risk assessment with matched zones.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude, buildingType]
              properties:
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                  example: 10.0889
                  description: GPS latitude (WGS84)
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
                  example: 76.3910
                  description: GPS longitude (WGS84)
                buildingType:
                  type: string
                  enum: [residential, commercial, industrial, institutional, agricultural]
                  example: residential
      responses:
        '200':
          description: Feasibility assessment result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      checkId: { type: integer }
                      overallRisk: { type: string, enum: [none, low, moderate, high, very_high, catastrophic] }
                      floodRisk:
                        type: object
                        properties:
                          found: { type: boolean }
                          count: { type: integer }
                          zones: { type: array }
                      landslideRisk:
                        type: object
                        properties:
                          found: { type: boolean }
                          count: { type: integer }
                          zones: { type: array }
                      coastalRisk:
                        type: object
                        properties:
                          found: { type: boolean }
                          count: { type: integer }
                          zones: { type: array }
                      seismicRisk:
                        type: object
                        properties:
                          found: { type: boolean }
                          count: { type: integer }
                          zones: { type: array }
                      historicProximity:
                        type: object
                        properties:
                          found: { type: boolean }
                          count: { type: integer }
                          events: { type: array }
        '400':
          description: Validation error
  /api/v1/feasibility/history:
    get:
      tags: [Feasibility]
      summary: Paginated history of feasibility checks
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          description: List of past checks with pagination metadata
  /api/v1/feasibility/{id}:
    get:
      tags: [Feasibility]
      summary: Retrieve a past feasibility check by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Feasibility check details
        '404':
          description: Check not found
...